Move lyric blueprint area into single layer.
